Signed-off-by: Paul Coombs <paulccoombs@gmail.com>
This commit reverts the check for renouncing guild rank (to 60cap) to the 75-era condition of 2 crafts being craftsman or higher.

Previously, the renounce was fixed to revert to 60cap (from the retail of 70), but the condition to activate the dialog was not.
If you have any 2 crafts at craftsman or higher, you should recieve the option from the guild master to reset your rank to journeyman.